% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm2.R
\name{algorithm2}
\alias{algorithm2}
\title{Algorithm 2: Updating the non clustering betas}
\usage{
algorithm2(Y, Z, mi_vec, lp, B, beta, D, gammas, phi, C)
}
\arguments{
\item{C}{}
}
\value{
Vector of length PK x L
}
\description{
Algorithm 2: Updating the non clustering betas
}
