% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm1.R
\name{beta_cluster}
\alias{beta_cluster}
\title{Group update beta}
\usage{
beta_cluster(y, Z, beta, lp, lp_minus, B, clusters, K, P, M)
}
\arguments{
\item{y}{#' A matrix of response values (e.g., counts), with \eqn{M} rows and \eqn{K} columns. Each row corresponds to a subject-time combination.}

\item{Z}{formatted external variable matrix with intercept column}

\item{beta}{Original unpenalized beta}

\item{lp}{clustering index}

\item{lp_minus}{non clustering indeces}

\item{B}{B spline basis}

\item{clusters}{identified clusters}

\item{K}{number of responses}

\item{P}{order + nknots}

\item{M}{total rows, timepoints times subjects}
}
\value{
beta matrix where betas are equal in each cluster
}
\description{
Group update beta
}
