% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm1.R
\name{beta_cluster_group_update}
\alias{beta_cluster_group_update}
\title{Title}
\usage{
beta_cluster_group_update(y, Z, beta, lp, lp_minus, B, clusters, K, P, M)
}
\arguments{
\item{M}{}
}
\description{
Title
}
