% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{beta_path}
\alias{beta_path}
\title{Plot the change in betas through the loop}
\usage{
beta_path(betas, K, B, Z, time)
}
\arguments{
\item{betas}{list of betas}
}
\value{
ggplot object
}
\description{
Plot the change in betas through the loop
}
