% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC_cluster_group}
\alias{BIC_cluster_group}
\title{BIC_cluster_group}
\usage{
BIC_cluster_group(
  y_hat_counts,
  y_counts,
  beta_group,
  clusters,
  M,
  K,
  nknots,
  order
)
}
\arguments{
\item{y_hat_counts}{Estimated y hat in counts}

\item{y_counts}{True Y counts}

\item{beta_group}{Re-fit group betas}

\item{clusters}{Clusters}

\item{M}{Total time points x samples}

\item{K}{Number of responses}

\item{nknots}{Number of knots}

\item{order}{Order of the B-spline basis}
}
\value{
List of BIC, first term, and second term
}
\description{
BIC_cluster_group
}
