% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly-package.R
\docType{package}
\name{cleverly-package}
\alias{cleverly-package}
\title{cleverly:}
\description{
cleverly is an R package implementing a novel clustering method for longitudinal microbiome data that incorporates external variables into the modeling framework. Unlike traditional clustering approaches, which often ignore time structure and external influences, cleverly uses a B-spline smoothing approach combined with Dirichlet-multinomial generalized estimating equations (GEEs) to account for compositional and temporal correlation in sequencing data. This enables clustering of microbial taxa based on either baseline behavior or their response to external variables, such as environmental or physiological conditions.
}
\section{Mypackage functions}{

The mypackage functions ...
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/empalmer/GEEclevR}
  \item Report bugs at \url{https://github.com/empalmer/GEEclevR/issues}
}

}
\author{
\strong{Maintainer}: Emily Palmer \email{palmerem@oregonstate.edu}

}
