% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly-package.R, R/cleverly.R
\name{cleverly}
\alias{cleverly}
\title{Cleverly: main function of the package}
\usage{
cleverly(
  Y,
  Z,
  subject_ids,
  time,
  lp = 0,
  response_type = "counts",
  cor_str = "IND",
  gammas,
  psi_min = 500,
  psi_max = 1500,
  npsi = 10,
  parralel = FALSE,
  nworkers = 2,
  tau = 0.005,
  theta = 300,
  C = 100,
  d = 2,
  nknots = 3,
  order = 3,
  epsilon_b = 0.001,
  epsilon_r = 0.001,
  epsilon_d = 0.001,
  epsilon_2 = 0.001,
  run_min = 3,
  max_outer_iter = 30,
  max_admm_iter = 100,
  max_2_iter = 100
)
}
\arguments{
\item{Y}{either a data frame or matrix of numeric response variables. Each response should be a separate column. Each row should be a separate subject/time combination. There should be M total rows. Must be ordered in time.}

\item{Z}{Matrix or data frame containing a column for each external variable. There should be M rows and L columns.}

\item{subject_ids}{either a vector of length(Y) or a column reference if Y is a data frame}

\item{time}{either a vector of length(Y) or a column reference if Y is a data frame. Must be numeric}

\item{lp}{either a numeric index of which external variable to cluster on, or the name of the column of Z that contains the clustering variable. Specify numeric 0 to cluster via baseline.}

\item{response_type}{Counts or continuous response}

\item{cor_str}{Type of correlation structure (IND, CON, AR1, CON-d, AR1-d)}

\item{gammas}{Vector of dimension L + 1 for penalizing the D matrix}

\item{psi_min}{Psi, or minimum psi if npsi >1}

\item{psi_max}{Maximum psi}

\item{npsi}{Number of psi values to test. Will create an equally spaced sequence from psi_min to psi_max and find the best (via BIC) choice.}

\item{parralel}{Use parralelization? Default is FALSE. If TRUE, uses future::plan(future::multisession) to run the algorithm in parallel.}

\item{nworkers}{If parallel, how many workers to use}

\item{tau}{MCP hyper parameter. Default is 8/100.}

\item{theta}{ADMM hyper parameter. Default is 300.}

\item{C}{Constant for determining the hessian change. Default is 100.}

\item{d}{Order for the difference matrix}

\item{nknots}{Number of knots for the B-spline basis}

\item{order}{Order of the B-spline basis}

\item{epsilon_b}{Tolerance for alg 1 convergence}

\item{epsilon_r}{Tolerance for ADMM convergence}

\item{epsilon_d}{Tolerance for ADMM convergence}

\item{epsilon_2}{Tolerance for convergence of algorithm 2}

\item{run_min}{minimum number of runs}

\item{max_outer_iter}{Number of iterations for the outer loop (Algorithm 1)}

\item{max_admm_iter}{Number of iterations for the clustering step (Algorithm 3)}

\item{max_2_iter}{Maximum number of iterations for algorithm 2 to run each loop (Algorithm 2)}
}
\value{
cluster" = res$clusters,

"possible_clusters" = res$all_clusters_psi,
"chosen_psi" = psis\link{best},
"y_hat_init" = res$y_hat_init,"y_hat" = res$y_hat)
}
\description{
Runs cleverly one psi for a range of psi values
}
