% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter.R
\name{cleverly_bestpsi}
\alias{cleverly_bestpsi}
\title{cleverly_bestpsi}
\usage{
cleverly_bestpsi(
  psi_min,
  psi_max,
  npsi,
  parralel = FALSE,
  Y,
  Z,
  time,
  lp = 0,
  response_type = "counts",
  cor_str = "IND",
  gammas,
  tau = 8/100,
  theta = 300,
  C = 10,
  d = 2,
  run_min = 3,
  nknots = 3,
  order = 3,
  epsilon_b = 0.001,
  epsilon_r = 0.001,
  epsilon_d = 0.001,
  max_outer_iter = 10,
  max_admm_iter = 100,
  max_2_iter = 100,
  epsilon_2 = 0.001
)
}
\arguments{
\item{psi_min}{minimum psi}

\item{psi_max}{maximum psi}

\item{npsi}{Number of psi values to test}

\item{parralel}{Run in parallel? T/F, if T, use all available cores}

\item{gammas}{Vector of dimension L + 1 for penalizing the D matrix}

\item{tau}{MCP hyper parameter.}

\item{theta}{ADMM hyper parameter.}

\item{run_min}{}

\item{max_admm_iter}{Max number of iterations for the ADMM loop}

\item{max_2_iter}{Maximum number of iterations for algorithm 2 to run each loop}
}
\description{
Run cleverly for a range of hyperparameter psi values.
}
