% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLR_cluster.R
\name{CLR_cluster}
\alias{CLR_cluster}
\title{CLR cluster}
\usage{
CLR_cluster(Y, Z, time, B, cluster_index, cluster_method = "kmeans", K, P, M)
}
\arguments{
\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{Z}{Full Z matrix (including intercept)}

\item{time}{Vector of numeric time values.}

\item{B}{B spline basis matrix}

\item{cluster_index}{Which column is clustering can only be 0 or 1}

\item{cluster_method}{One of kmeans, hclust, pam}

\item{K}{Number of responses}

\item{P}{number of B-spline coefficients (order + nknots)}

\item{M}{Number of samples times time points for each sample}
}
\description{
Note: only implemented for the case of 1 external variable.
}
