% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{correct_cluster_option}
\alias{correct_cluster_option}
\title{Get proportion correct cluster possible.}
\usage{
correct_cluster_option(res, true_cluster = rep(1:3, each = 4))
}
\arguments{
\item{res}{cleverly object}

\item{true_cluster}{true cluster membership}
}
\value{
Proportion of simulation runs the correct cluster was an option
}
\description{
Get proportion correct cluster possible.
}
