% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{Dirichlet.multinomial}
\alias{Dirichlet.multinomial}
\title{Simulate Dirichlet multinomial counts.}
\usage{
Dirichlet.multinomial(Y0, alpha)
}
\arguments{
\item{Y0}{Total sum of counts in DM}

\item{alpha}{}
}
\value{
DM counts
}
\description{
Simulate Dirichlet multinomial counts.
}
