% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_DM_counts}
\alias{generate_DM_counts}
\title{Simulate Dirichlet multinomial counts.}
\usage{
generate_DM_counts(Y0, alpha)
}
\arguments{
\item{Y0}{Total sum of counts in DM}

\item{alpha}{Vector of Dirichlet parameters}
}
\value{
DM counts
}
\description{
Simulate Dirichlet multinomial counts.
}
