% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_alpha_ij}
\alias{get_alpha_ij}
\title{Calculate alpha for all k for a single i and j.}
\usage{
get_alpha_ij(i, j, beta, Z, B, K, mi_vec)
}
\arguments{
\item{i}{subject index}

\item{j}{time index}

\item{beta}{beta matrix of dimension (P*K) x L}

\item{Z}{matrix of external variables of dimension (N x L)}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{K}{Number of parameters}

\item{mi}{Number of timepoints for subject i}
}
\value{
Vector of length K
}
\description{
Calculate alpha for all k for a single i and j.
}
