% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_B}
\alias{get_B}
\title{Create B spline basis matrix}
\usage{
get_B(time, order, nknots)
}
\arguments{
\item{time}{vector of time values for each subject/time}

\item{order}{Order of the B-spline basis}

\item{nknots}{Number of knots for the B-spline basis}
}
\value{
B spline basic matrix of dimension (M x P)
}
\description{
Create B spline basis matrix
}
