% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_B_ij}
\alias{get_B_ij}
\title{Get \eqn{B_ij}}
\usage{
get_B_ij(i, j, B, mi)
}
\arguments{
\item{i}{subject}

\item{j}{time index}

\item{B}{B-spline basis matrix}

\item{mi}{number of time points for subject i}
}
\value{
A vector of length P
}
\description{
Get \eqn{B_ij}
}
