% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_beta_kl}
\alias{get_beta_kl}
\title{Get \eqn{\beta_{kl}}}
\usage{
get_beta_kl(k, l, beta, P)
}
\arguments{
\item{k}{response index}

\item{l}{external variable index}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{P}{number of B-spline coefficients}
}
\value{
Vector of length P
}
\description{
Get \eqn{\beta_{kl}}
}
