% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Bij}
\alias{get_Bij}
\title{Get \eqn{B_ij}$}
\usage{
get_Bij(i, j, B, mi)
}
\arguments{
\item{i}{subject}

\item{j}{time index}

\item{B}{B-spline basis matrix}

\item{mi}{number of time points for subject i}
}
\description{
Get \eqn{B_ij}$
}
