% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm1.R
\name{get_D}
\alias{get_D}
\title{Construct \eqn{d}th-Order Difference Operator Matrix}
\usage{
get_D(K, d, order, nknots)
}
\arguments{
\item{K}{The number of response variables.}

\item{d}{The order of the difference operator (currently only \code{d = 2} is supported).}

\item{order}{The order of the B-spline basis functions.}

\item{nknots}{The number of internal knots used in the B-spline basis.}
}
\value{
A matrix \code{D} representing the second-order difference operator applied across response curves.
}
\description{
Returns the difference operator matrix \eqn{D} used to penalize differences in coefficient values across responses.
\strong{Note:} Currently implemented only for \eqn{d = 2}.
}
