% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{get_dHessian_il}
\alias{get_dHessian_il}
\title{Get hessian elements}
\usage{
get_dHessian_il(
  i,
  l,
  Y0,
  mi_vec,
  i_index,
  beta,
  Z,
  B,
  phi,
  Vi_inv,
  partials_il,
  alpha_i,
  K,
  P
)
}
\arguments{
\item{i}{subject index}

\item{l}{external variable index}

\item{Y0}{Vector of total count for each sample}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{i_index}{starting index of the ith subject in the data}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{phi}{Current value of overdispersion parameter}

\item{Vi_inv}{(optional) Vi inverse If supplied is faster}

\item{partials_il}{Partials vector for ith subject and lth external variable}

\item{alpha_i}{Vector of DM parameters for i}

\item{K}{Number of responses}

\item{P}{Number of B-spline coefficients (order + nknots)}
}
\value{
Matrix of dimension KP times KP
}
\description{
Get for for the ith subject deriving with respect to the lth external variable
}
