% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{get_gradient_l}
\alias{get_gradient_l}
\title{Title}
\usage{
get_gradient_l(Y, subject_ids, time_ids, l, phi, beta, Z, B)
}
\arguments{
\item{Y}{Matrix of dimension M x K}

\item{subject_ids}{numeric}

\item{time_ids}{numeric}

\item{l}{external variable index}

\item{phi}{variance parameter}

\item{K}{}
}
\value{
vector of length PK x 1
}
\description{
Title
}
