% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{get_Hessian_l}
\alias{get_Hessian_l}
\title{Get hessian with respect to the lth external variable}
\usage{
get_Hessian_l(
  l,
  Y0,
  mi_vec,
  i_index,
  beta,
  Z,
  B,
  phi,
  C,
  V_inv,
  partials_l,
  alpha,
  P,
  K
)
}
\arguments{
\item{l}{external variable index}

\item{Y0}{Vector of total count for each sample}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{i_index}{starting index of the ith subject in the data}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{phi}{Current value of overdispersion parameter}

\item{C}{Constant for determining the hessian change.}

\item{V_inv}{(optional) List of V inverses for each i. If supplied is faster}

\item{K}{Number of responses}
}
\value{
Matrix of dimension KP times KP
}
\description{
Get hessian with respect to the lth external variable
}
