% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_indeces}
\alias{get_indeces}
\title{Title}
\usage{
get_indeces(i, mi_vec)
}
\arguments{
\item{i}{subject index}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}
}
\value{
indeces for each ith subject repeated mi times
}
\description{
Title
}
