% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_knots}
\alias{get_knots}
\title{Get breaks for B-spline basis}
\usage{
get_knots(t, k, m)
}
\arguments{
\item{t}{timepoints}

\item{k}{order of B-spline}

\item{m}{number of knots}
}
\value{
vector of knot values
}
\description{
Get breaks for B-spline basis
}
