% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mi_vec}
\alias{get_mi_vec}
\title{Get data frame of mi values for each i}
\usage{
get_mi_vec(Y, subject_ids, time_ids)
}
\arguments{
\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{subject_ids}{either a vector of length(Y) or a column reference if Y is a data frame}

\item{time_ids}{either a vector of length(Y) or a column reference if Y is a data frame. Must be numeric}
}
\value{
data frame of mi and subject id
}
\description{
Get data frame of mi values for each i
}
