% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mu_i}
\alias{get_mu_i}
\title{Get mu_i mean vector for ith sample}
\usage{
get_mu_i(i, alpha, mi_vec, i_index, Y0, K)
}
\arguments{
\item{i}{subject index}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{i_index}{starting index of the ith subject in the data}

\item{Y0}{Vector of total count for each sample}

\item{K}{Number of responses}
}
\value{
Vector of length K*mi with each K first
}
\description{
Get mu_i mean vector for ith sample
}
