% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mu_ij}
\alias{get_mu_ij}
\title{Get the ijth values for mu}
\usage{
get_mu_ij(Y_ij0, alpha_ij, i, j, beta, Z, B, K, mi_vec)
}
\arguments{
\item{Y_ij0}{Total sum of counts across all K}

\item{alpha_ij}{Vector of DM parameters for i, j}

\item{i}{subject index}

\item{j}{time index}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{K}{Number of responses}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}
}
\value{
Vector of length K
}
\description{
Get the ijth values for mu
}
