% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{get_partials_l_list}
\alias{get_partials_l_list}
\title{Partials list (of matrices) with respect to given l}
\usage{
get_partials_l_list(Y0, l, mi_vec, i_index, beta, alpha, Z, B, P, K)
}
\arguments{
\item{Y0}{Vector of total count for each sample}

\item{l}{external variable index}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{i_index}{starting index of the ith subject in the data}

\item{beta}{List of beta values}

\item{alpha}{list of alpha that can be subsetted by i and j}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{P}{Number of B-spline coefficients (order + nknots)}

\item{K}{Number of responses}
}
\value{
list
}
\description{
Partials list (of matrices) with respect to given l
}
