% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{get_pearson_residual_i}
\alias{get_pearson_residual_i}
\title{Get the pearson residual for a given i}
\usage{
get_pearson_residual_i(Y, i, beta, Z, B, K, mi_vec)
}
\arguments{
\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{i}{subject index}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{K}{Number of responses}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{phi}{Current value of overdispersion parameter}
}
\value{
pearson residual vector for each i
}
\description{
Get the pearson residual for a given i
}
