% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm3.R
\name{get_r_norm}
\alias{get_r_norm}
\title{get_r_norm}
\usage{
get_r_norm(beta, v, Kappa, P, lp)
}
\arguments{
\item{Kappa}{keeps track of each possible response pair}

\item{lp}{}
}
\description{
Get the L2 norm of convergence criteria r
}
