% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_U_ij}
\alias{get_U_ij}
\title{Get term U_ij}
\usage{
get_U_ij(alpha_ij)
}
\arguments{
\item{alpha_ij}{Vector of DM parameters for i, j}
}
\value{
U ij part of partials/variance matrix
}
\description{
Get term U_ij
}
