% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_V_i}
\alias{get_V_i}
\title{Get Vi diagonal matrix of all js.}
\usage{
get_V_i(i, Y, Y_ij0, phi, beta, Z, B, K, mi)
}
\arguments{
\item{mi}{}
}
\value{
Diagonal matrix of Kmi x Kmi Each block is for a single j of block size K x K
}
\description{
Get Vi diagonal matrix of all js.
}
