% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_inverse.R
\name{get_V_ijj}
\alias{get_V_ijj}
\title{Get DM variance for ith subject jth timepoint}
\usage{
get_V_ijj(Y_ij0, phi, alpha_ij)
}
\arguments{
\item{Y_ij0}{Total sum of counts across all K}

\item{phi}{Current value of overdispersion parameter}

\item{alpha_ij}{Vector of DM parameters for i, j}
}
\value{
matrix of dimension K x K for a single j
}
\description{
Get DM variance for ith subject jth timepoint
}
