% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{get_Vi_inv}
\alias{get_Vi_inv}
\title{Get Vi inverse}
\usage{
get_Vi_inv(V_i, i, Y, mi_vec, phi, beta, Z, B, K)
}
\arguments{
\item{V_i}{}
}
\value{
Matrix of dimension Kmi x Kmi
}
\description{
Get Vi inverse
}
