% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Y_wrapper}
\alias{get_Y_wrapper}
\title{Get Yi for multiple user input options}
\usage{
get_Y_wrapper(Y, subject_ids, time_ids)
}
\arguments{
\item{Y}{either a data frame or matrix. Each response should be a separate column. Each row should be a separate subject/time combination. There should be M total rows.}

\item{subject_ids}{either a vector of length(Y) or a column reference if Y is a data frame}

\item{time_ids}{either a vector of length(Y) or a column reference if Y is a data frame}
}
\value{
Matrix of dimension mi x K
}
\description{
Users can either specify Y as a matrix or data frame.
}
\details{
The subject ids and time ids can either be references to columns in Y or external vectors that link the ids to the rows of Y
}
