% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{get_Yi_minus_mui}
\alias{get_Yi_minus_mui}
\title{Get the Yi minus mui term of the gradient}
\usage{
get_Yi_minus_mui(i, Y, mi_vec, beta, Z, B, K)
}
\arguments{
\item{mi_vec}{}
}
\value{
Vector of length Kmi
}
\description{
Get the Yi minus mui term of the gradient
}
