% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Yij0}
\alias{get_Yij0}
\title{Get \eqn{Y_{ij0}}$}
\usage{
get_Yij0(i, j, Y, mi)
}
\arguments{
\item{i}{subject}

\item{j}{time index}

\item{Y}{response}

\item{mi}{number of time points for subject i}
}
\description{
Get \eqn{Y_{ij0}}$
}
