% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Z_ijl}
\alias{get_Z_ijl}
\title{Get \eqn{Z_{ijl}}$}
\usage{
get_Z_ijl(i, j, l, Z)
}
\arguments{
\item{i}{subject}

\item{j}{time index}

\item{l}{external variable index}

\item{Z}{matrix of external variables}
}
\description{
Get \eqn{Z_{ijl}}$
}
