% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{initialize_beta}
\alias{initialize_beta}
\title{Initialize beta}
\usage{
initialize_beta(K, L, P)
}
\arguments{
\item{K}{Number of responses}

\item{L}{Number of external variables}

\item{P}{Number of B-spline coefficients (order + nknots)}
}
\value{
0 matrix of dimension KP x (L + 1)
}
\description{
Initialize beta
}
