% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm1.R
\name{initialize_beta_count}
\alias{initialize_beta_count}
\title{Initialize beta (for count data)}
\usage{
initialize_beta_count(K, L, P, B, Y, Z)
}
\arguments{
\item{K}{Number of responses}

\item{L}{Number of external variables}

\item{P}{Number of B-spline coefficients (order + nknots)}
}
\value{
0 matrix of dimension KP x (L + 1)
}
\description{
For each response independently, use the least squares estimate using the B-spline basis matrix and log Y
}
