% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_alg2_convergence}
\alias{plot_alg2_convergence}
\title{plot_alg2_convergence}
\usage{
plot_alg2_convergence(res)
}
\arguments{
\item{res}{Cleverly model object}
}
\value{
ggplot object
}
\description{
plot_alg2_convergence
}
