% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_BIC}
\alias{plot_BIC}
\title{Plot BIC values for a result.}
\usage{
plot_BIC(res, BIC_type = "BIC", psis)
}
\arguments{
\item{res}{cleverly object}

\item{BIC_type}{which version of BIC to plot}

\item{psis}{sequence of psis used}
}
\value{
ggplot object
}
\description{
Plot BIC values for a result.
}
