% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_cluster_differences}
\alias{plot_cluster_differences}
\title{Plot an alluvial plot to compare two cluster results}
\usage{
plot_cluster_differences(
  res1,
  res2,
  response_names = NULL,
  res_names = c("A", "B")
)
}
\arguments{
\item{res1}{cluster result 1}

\item{res2}{cluster result 2}

\item{response_names}{vector of response names}

\item{res_names}{vector of length two to name the cluster results}
}
\value{
ggplot object
}
\description{
Plot an alluvial plot to compare two cluster results
}
