% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_diagnostics.R
\name{plot_cluster_path}
\alias{plot_cluster_path}
\title{Title}
\usage{
plot_cluster_path(
  res,
  psi,
  tau,
  theta,
  gammas,
  max_admm_iter,
  max_outer_iter,
  duration
)
}
\arguments{
\item{duration}{}
}
\description{
Title
}
