% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clusters}
\usage{
plot_clusters(
  res,
  Z,
  response_names,
  order = "response",
  Z_type = "binary",
  y_type = "y_hat_baseline",
  lp_curve_only = T,
  nrow = 3,
  scales = "fixed",
  EV_color = "grey50"
)
}
\arguments{
\item{res}{cleverly model object}

\item{Z}{value of Z}

\item{response_names}{vector of names for each response}

\item{order}{"response" or "cluster" to have original ordering or cluster ordering of responses on the facet plots}

\item{Z_type}{binary or continuous}

\item{y_type}{which yhat to plot, either "y_hat_baseline", "y_hat_lp_group", or "slope"}

\item{lp_curve_only}{plot only the lp_curve_only? Has to be true for continuous, but for binary slope can be plotted}

\item{nrow}{number of rows for the facet plot}

\item{scales}{"free_y" or "fixed" to control scaling of the y axis}

\item{EV_color}{color for the EV line}
}
\value{
ggplot object
}
\description{
For binary EV
}
