% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_diagnostics.R
\name{plot_clusters_yhat}
\alias{plot_clusters_yhat}
\title{Plot clusters based on a yhat data frame}
\usage{
plot_clusters_yhat(yhat, chosen_cluster, K = 12)
}
\arguments{
\item{K}{Number of responses}
}
\description{
yhat should contain yhat, y, Z, and time
}
