% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_initial_fit}
\alias{plot_initial_fit}
\title{plot_initial_fit}
\usage{
plot_initial_fit(res, response_names = NULL, Z_col)
}
\arguments{
\item{res}{Cleverly model object}

\item{response_names}{Vector of response names.}

\item{Z_col}{Z column used}
}
\value{
ggplot object
}
\description{
plot_initial_fit
}
