% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_one_cluster}
\alias{plot_one_cluster}
\title{Title}
\usage{
plot_one_cluster(
  res,
  response_names,
  cluster_val,
  curve_type = "baseline",
  nrow = 3,
  scales = "fixed",
  EV_color = "grey50",
  Y_counts = NULL
)
}
\arguments{
\item{Y_counts}{}
}
\description{
Title
}
