% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{sim_data_same_base_different_slope}
\alias{sim_data_same_base_different_slope}
\title{Generating data simulation function}
\usage{
sim_data_same_base_different_slope(n, time, miss_n, cor_matrix, ranges)
}
\arguments{
\item{n}{n constant of number of individuals}

\item{time}{vector of time values for each subject/time}

\item{miss_n}{missingness}

\item{cor_matrix}{correlation matrix}

\item{ranges}{ranges}

\item{Z}{}
}
\value{
Simulated data frame
}
\description{
Generating data simulation function
}
