% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{sim_timepoints}
\alias{sim_timepoints}
\title{Simulate timepoints}
\usage{
sim_timepoints(n, maxt = 9, max_mi = 4)
}
\arguments{
\item{n}{number of subjects}

\item{maxt}{maximum time}

\item{max_mi}{maximum number of timepoints for each subject}
}
\value{
list
}
\description{
Get a collection of timepoints along the same range
}
