% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{sim_Y_ij}
\alias{sim_Y_ij}
\title{Simulate Yij counts}
\usage{
sim_Y_ij(i, j, beta, Z, B, Y_ij0, K, mi_vec)
}
\arguments{
\item{i}{subject index}

\item{j}{time index}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{Y_ij0}{Total sum of counts across all K}

\item{K}{Number of responses}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}
}
\value{
Y_i
}
\description{
Use Dirichlet Multinomial distribution to simulate counts
}
