% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{sim_Z}
\alias{sim_Z}
\title{Simulate 2 external variables, one continuous and one binary}
\usage{
sim_Z(mi_vec)
}
\arguments{
\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}
}
\value{
Formatted Z matrix
}
\description{
Simulate 2 external variables, one continuous and one binary
}
