% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{sim_Z_longitudinal}
\alias{sim_Z_longitudinal}
\title{Simulate data}
\usage{
sim_Z_longitudinal(
  n = 20,
  range_start = 5000,
  range_end = 20000,
  nknots = 3,
  K = 12,
  order = 3,
  user_var = 1000,
  cor_str,
  rho = 0.4,
  miss_p = 0.6,
  prob1 = 0.4,
  slope_base = "cluster_base_alldiff_slope"
)
}
\arguments{
\item{n}{Number of subjects}

\item{range_start}{Starting count range for the Dirichlet multinomial}

\item{range_end}{Ending count range for the Dirichlet multinomial}

\item{nknots}{Number of knots}

\item{K}{Number of responses}

\item{order}{Order of the B-spline basis}

\item{user_var}{User supplied variance}

\item{cor_str}{Type of correlation structure (IND, CON, AR1)}

\item{rho}{Correlation parameter}

\item{miss_p}{proportion of missing samples}

\item{prob1}{Probability of the binary external variable being 1}

\item{slope_base}{type of slope/intercept clustering curves to generate}
}
\value{
data Matrix with columns time, individual, capture number, totaln, counts
}
\description{
Use Chenyang's setup to simulate count data with 3 clusters,
one external variable binary with prob prob1 of being 1
}
