% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm3.R
\name{update_beta_admm}
\alias{update_beta_admm}
\title{Update beta_lp in the ADMM algorithm}
\usage{
update_beta_admm(
  Y,
  beta,
  lp,
  v,
  lambda,
  theta,
  gammas,
  D,
  A,
  Z,
  B,
  phi,
  C,
  mi_vec
)
}
\arguments{
\item{mi_vec}{}
}
\description{
Update beta_lp in the ADMM algorithm
}
