% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{visualize_final_fit}
\alias{visualize_final_fit}
\title{Plot the final fit}
\usage{
visualize_final_fit(Y, res)
}
\arguments{
\item{Y}{Input matrix Matrix of counts Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{res}{cleverly model output object}
}
\value{
ggplot object
}
\description{
Plot the final fit of the model (relative abundance scale) with the data points (relative abundance scale)
}
