% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm1.R
\name{algorithm1}
\alias{algorithm1}
\title{Algorithm 1}
\usage{
algorithm1(Y, Z, subject_ids, time, lp, gamma, d, nknots, order, tol, smax)
}
\arguments{
\item{Y}{Matrix Each response should be a separate column. Each row should be a separate subject/time combination. There should be M total rows.}

\item{Z}{matrix that starts with a column of 1s. Of dimension M x L + 1}

\item{subject_ids}{vector of numeric ids for each subject}

\item{time}{vector of numeric ids}

\item{lp}{clustering index}

\item{gamma}{Vector of length (L + 1) of penalization hyper parameters}

\item{d}{Order for the difference matrix}

\item{nknots}{Number of knots for the B-spline basis}

\item{order}{Order of the B-spline basis}

\item{tol}{Tolerance for convergence}

\item{smax}{Maximum number of iterations}
}
\description{
Algorithm 1
}
