% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm1.R
\name{algorithm1}
\alias{algorithm1}
\title{Algorithm 1}
\usage{
algorithm1(
  Y,
  Z,
  time,
  mi_vec,
  lp,
  gammas,
  psi,
  tau,
  theta,
  C,
  d,
  nknots,
  order,
  epsilon_b,
  epsilon_r,
  epsilon_d,
  max_outer_iter,
  max_admm_iter
)
}
\arguments{
\item{Y}{Matrix of counts Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{time}{vector of time values for each subject/time}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{lp}{clustering index (integer between 0 and L)}

\item{gammas}{Vector of dimension L + 1 for penalizing the D matrix}

\item{psi}{Hyperparameter for clustering penalty (larger drives pairwise differences to zero)}

\item{tau}{MCP hyper parameter.}

\item{theta}{ADMM hyper parameter.}

\item{C}{Constant for determining the hessian change.}

\item{d}{Order for the difference matrix}

\item{nknots}{Number of knots for the B-spline basis}

\item{order}{Order of the B-spline basis}

\item{epsilon_b}{Tolerance for alg 1 convergence}

\item{epsilon_r}{Tolerance for ADMM convergence}

\item{epsilon_d}{Tolerance for ADMM convergence}

\item{max_outer_iter}{Max number of iterations for the outer loop (Algorithm 1)}

\item{max_admm_iter}{Max number of iterations for the ADMM loop}
}
\value{
List of beta, clusters, y_hat, v, admm_diffs, admm_beta_list, loop_list_beta, loop_list_diff, phis_list
}
\description{
Algorithm 1
}
