% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm2.R
\name{algorithm2}
\alias{algorithm2}
\title{Algorithm 2: Updating the non clustering betas}
\usage{
algorithm2(
  Y,
  Y0,
  Z,
  mi_vec,
  i_index,
  lp,
  B,
  beta,
  D,
  gammas,
  phi,
  C,
  max_2_iter,
  epsilon_2,
  time,
  s,
  L,
  P,
  K,
  M,
  cor_str
)
}
\arguments{
\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{Y0}{Vector of total count for each sample}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{i_index}{starting index of the ith subject in the data}

\item{lp}{clustering index (integer between 0 and L)}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{D}{Matrix of dth order weighted difference operator}

\item{gammas}{Vector of dimension L + 1 for penalizing the D matrix}

\item{phi}{Current value of overdispersion parameter}

\item{C}{Constant for determining the hessian change.}

\item{max_2_iter}{Maximum number of iterations for algorithm 2 to run each loop}

\item{L}{Number of external variables}

\item{K}{Number of responses}
}
\value{
Vector of length PK x L
}
\description{
Algorithm 2: Updating the non clustering betas
}
