% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm3.R
\name{algorithm3}
\alias{algorithm3}
\title{ADMM algorithm for updating the clustering beta}
\usage{
algorithm3(
  Y,
  Z,
  lp,
  B,
  beta,
  A,
  P,
  Kappa,
  mi_vec,
  gammas,
  tau,
  theta,
  psi,
  phi,
  max_admm_iter = 100
)
}
\arguments{
\item{max_admm_iter}{}
}
\description{
ADMM algorithm for updating the clustering beta
}
