% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm1.R
\name{BIC_cluster_group}
\alias{BIC_cluster_group}
\title{BIC_cluster_group}
\usage{
BIC_cluster_group(
  y_hat_counts,
  y_counts,
  beta_group,
  clusters,
  M,
  K,
  nknots,
  order
)
}
\arguments{
\item{order}{}
}
\description{
BIC_cluster_group
}
