% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly-package.R
\docType{package}
\name{cleverly-package}
\alias{cleverly-package}
\title{cleverly: A package for computating the notorious bar statistic.}
\description{
The mypackage package provides three categories of important functions:
foo, bar and baz.
}
\section{Mypackage functions}{

The mypackage functions ...
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/empalmer/GEEclevR}
  \item Report bugs at \url{https://github.com/empalmer/GEEclevR/issues}
}

}
\author{
\strong{Maintainer}: Emily Palmer \email{palmerem@oregonstate.edu}

}
