% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly-package.R, R/cleverly.R
\name{cleverly}
\alias{cleverly}
\title{cleverly}
\usage{
cleverly(
  Y,
  Z,
  subject_ids,
  time,
  lp = 0,
  response_type = "counts",
  cor_str = "IND",
  gammas,
  psi,
  tau = 8/100,
  theta = 300,
  C = 10,
  d = 2,
  run_min = 3,
  nknots = 3,
  order = 3,
  epsilon_b = 0.001,
  epsilon_r = 0.001,
  epsilon_d = 0.001,
  max_outer_iter = 10,
  max_admm_iter = 100,
  max_2_iter = 100,
  epsilon_2 = 0.001
)
}
\arguments{
\item{Y}{either a data frame or matrix of numeric response variables. Each response should be a separate column. Each row should be a separate subject/time combination. There should be M total rows. Must be ordered in time.}

\item{Z}{Matrix or data frame containing a column for each external variable. There should be M rows and L columns.}

\item{subject_ids}{either a vector of length(Y) or a column reference if Y is a data frame}

\item{time}{either a vector of length(Y) or a column reference if Y is a data frame. Must be numeric}

\item{lp}{either a numeric index of which external variable to cluster on, or the name of the column of Z that contains the clustering variable. Specify numeric 0 to cluster via baseline.}

\item{response_type}{Counts or continuous response?}

\item{gammas}{Vector of dimension L + 1 for penalizing the D matrix}

\item{psi}{Hyperparameter for clustering penalty (larger drives pairwise differences to zero)}

\item{tau}{MCP hyper parameter. Default is 8/100.}

\item{theta}{ADMM hyper parameter. Default is 300.}

\item{C}{Constant for determining the hessian change. Default is 10.}

\item{d}{Order for the difference matrix}

\item{run_min}{}

\item{nknots}{Number of knots for the B-spline basis}

\item{order}{Order of the B-spline basis}

\item{max_outer_iter}{Number of iterations for the outer loop (Algorithm 1)}

\item{max_admm_iter}{Number of iterations for the clustering step (Algorithm 3)}

\item{max_2_iter}{Maximum number of iterations for algorithm 2 to run each loop}
}
\description{
Overall user interface for this algorithm
Formats data and calls \code{algorithm1()} to run the algorithm
}
