% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter.R
\name{cleverly_bestparam}
\alias{cleverly_bestparam}
\title{Title}
\usage{
cleverly_bestparam(
  param_grid,
  parralel = FALSE,
  Y,
  Z,
  time,
  lp = 0,
  response_type = "counts",
  cor_str = "IND",
  gamma_length,
  theta = 300,
  C = 10,
  d = 2,
  nknots = 3,
  order = 3,
  epsilon_b = 0.001,
  epsilon_r = 0.001,
  epsilon_d = 0.001,
  max_outer_iter = 10,
  max_admm_iter = 100,
  max_2_iter = 100,
  epsilon_2 = 0.001
)
}
\arguments{
\item{epsilon_2}{}
}
\description{
Title
}
