% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLR_cluster.R
\name{CLR_cluster}
\alias{CLR_cluster}
\title{CLR cluster}
\usage{
CLR_cluster(Y, Z, time, B, lp, K, P, M)
}
\arguments{
\item{M}{}
}
\description{
Note: only implemented for the case of 1 external variable.
}
