% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{cor_user}
\alias{cor_user}
\title{User defined correlation structure}
\usage{
cor_user(mi, user_var, cor_str, rho)
}
\arguments{
\item{mi}{number of time points for ith sample}

\item{user_var}{User supplied variance}

\item{cor_str}{specified correaltion structure (of IND, CON-d, or AR1d)}

\item{rho}{rho correlation parameter}
}
\value{
User defined correlation matrix (for a single sample i)
}
\description{
User defined correlation structure
}
