% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{cor_user}
\alias{cor_user}
\title{User defined correlation structure}
\usage{
cor_user(mi, user_var, cor_str, rho)
}
\arguments{
\item{mi}{number of timepoints for ith sample}

\item{rho}{rho?}

\item{structure}{Type of correlation 1 for (?) 2 for (?) 1: compound symmetry, 2: autoregressive, 3: indepdendent}
}
\value{
User defined correlation matrix.
}
\description{
User defined correlation structure
}
