% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_Z}
\alias{format_Z}
\title{Format Z}
\usage{
format_Z(Z)
}
\arguments{
\item{Z}{A matrix or data frame with columns of external variables for each subject/time}
}
\value{
A matrix with a column of 1s representing L = 0, and values for the other extenral variables
}
\description{
Format Z
}
