% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_data_longitudinal_compositional}
\alias{generate_data_longitudinal_compositional}
\title{Generating data simulation function}
\usage{
generate_data_longitudinal_compositional(
  n,
  time,
  miss_n,
  cor_matrix,
  ranges,
  curveparamenter
)
}
\arguments{
\item{n}{n constant of number of individuals}

\item{time}{vector of time values for each subject/time}

\item{miss_n}{missingness}

\item{cor_matrix}{correlation matrix}

\item{ranges}{ranges}

\item{curveparamenter}{alphas for the Dirichlet multinomial}
}
\value{
Simulated data frame
}
\description{
Generating data simulation function
}
