% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_data_longitudinal_compositionalZ}
\alias{generate_data_longitudinal_compositionalZ}
\title{Generating data simulation function}
\usage{
generate_data_longitudinal_compositionalZ(n, time, miss_n, cor_matrix, ranges)
}
\arguments{
\item{n}{n constant of number of individuals}

\item{time}{vector of time values for each subject/time}

\item{miss_n}{missingness}

\item{cor_matrix}{correlation matrix}

\item{ranges}{ranges}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}
}
\value{
Simulated data frame
}
\description{
Generating data simulation function
}
