% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm1.R
\name{get_A}
\alias{get_A}
\title{Get A matrix}
\usage{
get_A(Kappa, K, P)
}
\arguments{
\item{Kappa}{keeps track of each possible response pair}

\item{K}{Number of responses}

\item{P}{Number of B-spline coefficients (order + nknots)}
}
\value{
A matrix
}
\description{
This is the matrix that keeps track of the pairwise differences,
}
