% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_alpha_ij}
\alias{get_alpha_ij}
\title{Calculate alpha for all k for a single i and j.}
\usage{
get_alpha_ij(i, j, beta_ks, Z, B, K, i_index, L, P)
}
\arguments{
\item{i}{subject index}

\item{j}{time index}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{K}{Number of responses}

\item{i_index}{starting index of the ith subject in the data}

\item{L}{Number of external variables}

\item{P}{}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}
}
\value{
Vector of length K
}
\description{
Calculate alpha for all k for a single i and j.
}
