% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_alpha_ijk}
\alias{get_alpha_ijk}
\title{Calculate alpha for a single \eqn{i = 1, j = 1, k = 1} from \eqn{\beta}}
\usage{
get_alpha_ijk(i, j, k, beta, Z, B, mi)
}
\arguments{
\item{i}{subject index}

\item{j}{time index}

\item{k}{response index}

\item{beta}{matrix of dimension (P*K) x L}

\item{Z}{matrix of external variables of dimension N x L}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{mi}{Number of timepoints for subject i}
}
\value{
Numeric (1x1)
}
\description{
Calculate alpha for a single \eqn{i = 1, j = 1, k = 1} from \eqn{\beta}
}
