% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_alpha_list}
\alias{get_alpha_list}
\title{Title}
\usage{
get_alpha_list(beta, Z, B, K, i_index, mi_vec, L, P)
}
\arguments{
\item{K}{Number of responses}

\item{i_index}{starting index of the ith subject in the data}

\item{L}{Number of external variables}

\item{P}{}
}
\description{
Title
}
