% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_B_ij}
\alias{get_B_ij}
\title{Get \eqn{B_ij}}
\usage{
get_B_ij(i, j, B, i_index)
}
\arguments{
\item{i}{subject index}

\item{j}{time index}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}
}
\value{
A vector of length P
}
\description{
Get \eqn{B_ij}
}
