% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{get_cluster_diagnostics}
\alias{get_cluster_diagnostics}
\title{Look at clustering results if true clusters are known}
\usage{
get_cluster_diagnostics(res, true_cluster)
}
\arguments{
\item{res}{model object from cleverly}

\item{true_cluster}{vector with the true cluster membership}
}
\value{
res model with additional list item diagnostics that contains clustering results of rand, adj.rand, jaccard, CER, and nclust
}
\description{
Look at clustering results if true clusters are known
}
