% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm1.R
\name{get_clusters}
\alias{get_clusters}
\title{Get cluster membership}
\usage{
get_clusters(v, K, P)
}
\arguments{
\item{v}{final v (pairwise differences) from algorithm 1}

\item{K}{Number of responses}

\item{P}{Number of B-spline coefficients (order + nknots)}
}
\value{
Cluster list information, indeces and number of clusters
}
\description{
Get cluster membership
}
