% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_D}
\alias{get_D}
\title{Return dth order differnece operator matrix}
\usage{
get_D(K, d, order, nknots)
}
\arguments{
\item{K}{Number of responses}

\item{d}{Order of the difference operator}

\item{order}{Order of the B-spline basis}

\item{nknots}{Number of knots for the B-spline basis}
}
\value{
D matrix
}
\description{
Note that this only works for d = 2
}
