% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm3.R
\name{get_d_norm}
\alias{get_d_norm}
\title{Title}
\usage{
get_d_norm(v_new, v, theta, Kappa, K, P)
}
\arguments{
\item{theta}{ADMM hyper parameter.}

\item{Kappa}{keeps track of each possible response pair}

\item{P}{}
}
\description{
Title
}
