% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{get_gradient_il}
\alias{get_gradient_il}
\title{Get gradient for a given i, l}
\usage{
get_gradient_il(
  i,
  l,
  Y,
  Y0,
  mi_vec,
  i_index,
  phi,
  beta,
  alpha,
  Z,
  B,
  Vi_inv,
  partials_il,
  K
)
}
\arguments{
\item{i}{subject index}

\item{l}{external variable index}

\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{Y0}{Vector of total count for each sample}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{i_index}{starting index of the ith subject in the data}

\item{phi}{Current value of overdispersion parameter}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{alpha}{list of alpha that can be subsetted by i and j}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{Vi_inv}{Vi matrix inverse for subject i}

\item{partials_il}{Partials vector for ith subject and lth external variable}

\item{K}{Number of responses}
}
\value{
Vector of length PK
}
\description{
Get gradient for a given i, l
}
