% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{get_gradient_l}
\alias{get_gradient_l}
\title{Get the gradient for a given l}
\usage{
get_gradient_l(
  Y,
  Y0,
  mi_vec,
  i_index,
  l,
  phi,
  beta,
  alpha,
  Z,
  B,
  V_inv,
  partials_l,
  K,
  P
)
}
\arguments{
\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{Y0}{Vector of total count for each sample}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{i_index}{starting index of the ith subject in the data}

\item{l}{external variable index}

\item{phi}{Current value of overdispersion parameter}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{alpha}{list of alpha that can be subsetted by i and j}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{V_inv}{V inverse}

\item{partials_l}{List of partials vectors for each i}

\item{K}{Number of responses}

\item{P}{Number of B-spline coefficients (order + nknots)}
}
\value{
vector of length PK x 1
}
\description{
Get the gradient for a given l
}
