% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{get_gradient_l}
\alias{get_gradient_l}
\title{Get the gradient for a given l}
\usage{
get_gradient_l(Y, mi_vec, l, phi, beta, Z, B)
}
\arguments{
\item{Y}{Matrix of dimension M x K}

\item{mi_vec}{counts of length n}

\item{l}{external variable index}

\item{phi}{variance parameter}

\item{is}{numeric of length M}

\item{K}{}
}
\value{
vector of length PK x 1
}
\description{
Get the gradient for a given l
}
