% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm1.R
\name{get_knots}
\alias{get_knots}
\title{Get Breaks for B-Spline Basis}
\usage{
get_knots(t, k, m)
}
\arguments{
\item{t}{A numeric vector of timepoints over which the B-spline basis will be constructed.}

\item{k}{The order of the B-spline}

\item{m}{The number of internal knots to include.}
}
\value{
A numeric vector of knot positions, including boundary and internal knots, suitable for use in B-spline basis construction.
}
\description{
Computes the breakpoints (knots) needed to construct a B-spline basis given a vector of timepoints, the spline order, and the desired number of internal knots.
}
