% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mi_vec}
\alias{get_mi_vec}
\title{Get data frame of mi values for each i}
\usage{
get_mi_vec(Y, subject_ids, time_ids)
}
\arguments{
\item{time_ids}{}
}
\description{
Get data frame of mi values for each i
}
