% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mu_i}
\alias{get_mu_i}
\title{Get mu_i mean vector for ith sample}
\usage{
get_mu_i(i, mis, Y, beta, Z, B, K)
}
\arguments{
\item{K}{}
}
\value{
Vector of length K*mi with each K first
}
\description{
Get mu_i mean vector for ith sample
}
