% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mu_ij}
\alias{get_mu_ij}
\title{Get the ijth values for mu}
\usage{
get_mu_ij(Y_ij0, alpha_ij, i, j, beta, Z, B, K, mi)
}
\arguments{
\item{alpha_ij}{}
}
\value{
Vector of length K
}
\description{
Get the ijth values for mu
}
