% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mu_ij}
\alias{get_mu_ij}
\title{Get the ijth values for mu}
\usage{
get_mu_ij(Y_ij0, alpha_ij)
}
\arguments{
\item{Y_ij0}{Total sum of counts across all K}

\item{alpha_ij}{Vector of DM parameters for i, j}
}
\value{
Vector of length K
}
\description{
Get the ijth values for mu
}
