% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{get_partials_ijl}
\alias{get_partials_ijl}
\title{Get partials matrix for i, j, l}
\usage{
get_partials_ijl(i, j, l, mi, K, Y, Z, B, beta)
}
\arguments{
\item{beta}{}
}
\value{
Matrix of dimension PK x K
}
\description{
Get partials matrix for i, j, l
}
