% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{get_partials_ijl}
\alias{get_partials_ijl}
\title{Get partials matrix for i, j, l}
\usage{
get_partials_ijl(i, j, l, mi_vec, i_index, Y0, Z, B, beta, alpha_ij)
}
\arguments{
\item{i}{subject index}

\item{j}{time index}

\item{l}{external variable index}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{alpha_ij}{}
}
\value{
Matrix of dimension PK x K
}
\description{
Get partials matrix for i, j, l
}
