% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{get_partials_il}
\alias{get_partials_il}
\title{Get full partials matrix for a given i, l}
\usage{
get_partials_il(i, l, Y, Z, B, beta, mi_vec)
}
\arguments{
\item{i}{subject index}

\item{l}{external variable index}

\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}
}
\value{
Matrix of dimension KP x Kmi
}
\description{
Get full partials matrix for a given i, l
}
