% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{get_partials_il}
\alias{get_partials_il}
\title{Get full partials matrix for a given i, l}
\usage{
get_partials_il(i, l, Y, Z, B, beta, mis)
}
\arguments{
\item{P}{}
}
\description{
Get full partials matrix for a given i, l
}
