% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{get_performance_summary}
\alias{get_performance_summary}
\title{get performance summary table}
\usage{
get_performance_summary(list, output)
}
\arguments{
\item{list}{simulation list}

\item{output}{output type of "R", "html", or "latex"}
}
\description{
get performance summary table
}
