% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_parameters.R
\name{get_phi}
\alias{get_phi}
\title{Phi}
\usage{
get_phi(pearson_residuals, K, M, L, P, phi_old)
}
\arguments{
\item{pearson_residuals}{list of pearson residuals for each i}

\item{K}{Number of responses}

\item{M}{Number of samples times timepoints for each sample}

\item{L}{Number of external variables}

\item{P}{Number of bspline basis parameters (order + nknots)}
}
\value{
scalar phi
}
\description{
Phi
}
