% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_inverse.R
\name{get_rho}
\alias{get_rho}
\title{Get rho}
\usage{
get_rho(pearson_residuals, phi, K, mi_vec, M, cor_str, cor_blocks, j1_j2_list)
}
\arguments{
\item{phi}{}

\item{K}{Number of responses}

\item{mi_vec}{vector of the number of time points for each sample. Of length n}

\item{M}{Number of subjects times time points for each subject}

\item{cor_str}{Type of correlation structure (IND, CON, AR1)}
}
\value{
Numeric estimate of rho
}
\description{
Get rho
}
