% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_U_ij}
\alias{get_U_ij}
\title{Get term U_ij}
\usage{
get_U_ij(alpha_ij, i, j, beta, Z, B, K, i_index)
}
\arguments{
\item{alpha_ij}{Vector of DM parameters for i, j}

\item{i}{subject index}

\item{j}{time index}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{K}{Number of responses}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}
}
\value{
U ij part of partials/variance matrix
}
\description{
Get term U_ij
}
