% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_inverse.R
\name{get_V_i}
\alias{get_V_i}
\title{Get Vi diagonal matrix of all js.}
\usage{
get_V_i(i, Y, Y0, phi, beta, Z, B, K, mi_vec, i_index, L, P)
}
\arguments{
\item{i}{subject index}

\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{Y0}{Vector of total count for each sample}

\item{phi}{Current value of overdispersion parameter}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{K}{Number of responses}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{i_index}{starting index of the ith subject in the data}

\item{L}{Number of external variables}

\item{P}{Number of B-spline coefficients (order + nknots)}
}
\value{
Diagonal matrix of Kmi x Kmi Each block is for a single j of block size K x K
}
\description{
Get Vi diagonal matrix of all js.
}
