% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_V_ijj}
\alias{get_V_ijj}
\title{Get DM variance for ith subject jth timepoint}
\usage{
get_V_ijj(Y_ij0, phi, alpha_ij, i, j, beta, Z, B, K, mi)
}
\arguments{
\item{mi}{}
}
\value{
matrix of dimension K x K for a single j
}
\description{
Get DM variance for ith subject jth timepoint
}
