% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_inverse.R
\name{get_V_inv}
\alias{get_V_inv}
\title{Get overal V inverse}
\usage{
get_V_inv(Y, Y0, alpha, mi_vec, i_index, phi, beta, Z, B, K, cor_str, rho_cor)
}
\arguments{
\item{Y0}{}
}
\description{
Get overal V inverse
}
