% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_inverse.R
\name{get_V_inv}
\alias{get_V_inv}
\title{Get overall V inverse}
\usage{
get_V_inv(Y, Y0, alpha, mi_vec, i_index, phi, beta, Z, B, K, cor_str, rho_cor)
}
\arguments{
\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{Y0}{Vector of total count for each sample}

\item{alpha}{list of alpha that can be subsetted by i and j}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{i_index}{starting index of the ith subject in the data}

\item{phi}{Current value of overdispersion parameter}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{K}{Number of responses}
}
\value{
V inverse list
}
\description{
for all i
}
