% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_inverse.R
\name{get_Vi_inv}
\alias{get_Vi_inv}
\title{Get Vi inverse}
\usage{
get_Vi_inv(i, Y, mi_vec, i_index, phi, beta, Z, B, K)
}
\arguments{
\item{i}{subject index}

\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{phi}{Current value of overdispersion parameter}

\item{beta}{matrix of beta (or beta hat) of dimension (P*K) x L}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{B}{B spline basis matrix of dimension (N x P)}

\item{K}{Number of responses}

\item{V_i}{}
}
\value{
Matrix of dimension Kmi x Kmi
}
\description{
Get Vi inverse
}
