% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Y_i_mat}
\alias{get_Y_i_mat}
\title{Get \eqn{Y_i}}
\usage{
get_Y_i_mat(i, mi_vec, Y)
}
\arguments{
\item{i}{subject index}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}
}
\value{
Matrix of dimension mi x K
}
\description{
Get \eqn{Y_i}
}
