% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Y_i_vec}
\alias{get_Y_i_vec}
\title{Get Yi as a vector}
\usage{
get_Y_i_vec(i, mi, Y, Y_mat)
}
\arguments{
\item{Y_mat}{}
}
\value{
Vector of length K * mi
}
\description{
different k responses are listed after each other. so it goes
y(i,j=1,k=1), ... yi(j = mi, k = 1), yi(j = 1, k = 2)...
NOPE! Other way around...
}
