% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Y_ij0}
\alias{get_Y_ij0}
\title{Get \eqn{Y_{ij0}}}
\usage{
get_Y_ij0(i, j, Y, mi)
}
\arguments{
\item{i}{subject}

\item{j}{time index}

\item{Y}{response}

\item{mi}{number of time points for subject i}
}
\value{
Scalar of the total sum constraint for a given i, j
}
\description{
Get \eqn{Y_{ij0}}
}
