% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Yi_minus_mui}
\alias{get_Yi_minus_mui}
\title{Get the Yi minus mui term of the gradient}
\usage{
get_Yi_minus_mui(i, Y, Y0, alpha, mi_vec, i_index, K)
}
\arguments{
\item{i}{subject index}

\item{Y}{Matrix of counts. Each response should be a separate column (K). Each row should be a separate subject/time combination. There should be M total rows.}

\item{Y0}{Vector of total count for each sample}

\item{alpha}{list of alpha that can be subsetted by i and j}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}

\item{i_index}{starting index of the ith subject in the data}

\item{K}{Number of responses}
}
\value{
Vector of length Kmi
}
\description{
Get the Yi minus mui term of the gradient
}
