% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Z_ijl}
\alias{get_Z_ijl}
\title{Get \eqn{Z_{ijl}}}
\usage{
get_Z_ijl(i, j, l, Z, mi_vec)
}
\arguments{
\item{i}{subject index}

\item{j}{time index}

\item{l}{external variable index}

\item{Z}{matrix of external variables of dimension N x L}
}
\value{
Scalar for the l-th external variable for subject i at time j
}
\description{
Get \eqn{Z_{ijl}}
}
