% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Z_ijl}
\alias{get_Z_ijl}
\title{Get \eqn{Z_{ijl}}}
\usage{
get_Z_ijl(i, j, l, Z, mi_vec)
}
\arguments{
\item{i}{subject index}

\item{j}{time index}

\item{l}{external variable index}

\item{Z}{Matrix that starts with a column of 1s. Of dimension M x (L + 1) that contains the external variable values for each subject/time and is 1 for l = 0. In the case that there are no external variables this is a matrix with one column of 1s.}

\item{mi_vec}{vector of the number of timepoints for each sample. Of length n}
}
\value{
Scalar for the l-th external variable for subject i at time j
}
\description{
Get \eqn{Z_{ijl}}
}
