% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_diagnostics.R
\name{plot_cluster_path}
\alias{plot_cluster_path}
\title{plot_cluster_path}
\usage{
plot_cluster_path(
  res,
  psi,
  tau,
  theta,
  gammas,
  max_admm_iter,
  max_outer_iter,
  duration
)
}
\arguments{
\item{res}{Cleverly model object}

\item{psi}{Hyperparameter for clustering penalty (larger drives pairwise differences to zero)}

\item{tau}{MCP hyper parameter.}

\item{theta}{ADMM hyper parameter.}

\item{gammas}{Vector of dimension L + 1 for penalizing the D matrix}

\item{max_admm_iter}{Max number of iterations for the ADMM loop}

\item{max_outer_iter}{Max number of iterations for the outer loop (Algorithm 1)}

\item{duration}{}
}
\value{
ggplot object
}
\description{
plot_cluster_path
}
