% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_diagnostics.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Title}
\usage{
plot_clusters(res, K, tau, psi, gammas, theta, max_admm_iter, max_outer_iter)
}
\arguments{
\item{max_outer_iter}{}
}
\description{
Title
}
