% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clusters}
\usage{
plot_clusters(
  res,
  response_names,
  order = "response",
  Z_type = "binary",
  curve_type = "baseline",
  nrow = 3,
  scales = "fixed",
  EV_color = "grey50",
  Y_counts
)
}
\arguments{
\item{res}{cleverly model object}

\item{response_names}{vector of names for each response}

\item{order}{"response" or "cluster" to have original ordering or cluster ordering of responses on the facet plots}

\item{Z_type}{binary or continuous}

\item{curve_type}{= baseline, refit_baseline, or slope}

\item{nrow}{number of rows for the facet plot}

\item{scales}{"free_y" or "fixed" to control scaling of the y axis}

\item{EV_color}{color for the EV line}

\item{Y_counts}{need if curve_type = "slope}
}
\value{
ggplot object
}
\description{
For binary EV
}
