% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clusters}
\usage{
plot_clusters(
  res,
  Z,
  response_names,
  scales = "free_y",
  EV_color = "grey50",
  order = "response",
  nrow = 3,
  Z_type = "binary",
  y_type = "y_hat_baseline",
  baseline = T
)
}
\arguments{
\item{res}{cleverly model object}

\item{Z}{value of Z}

\item{response_names}{vector of names for each response}

\item{scales}{"free_y" or "fixed" to control scaling of the y axis}

\item{EV_color}{color for the EV line}

\item{order}{"response" or "cluster" to have original ordering or cluster ordering of responses on the facet plots}

\item{nrow}{number of rows for the facet plot}

\item{Z_type}{binary or continuous}

\item{baseline}{plot only baseline? Has to be true for continuous, but for binary slope can be plotted}
}
\value{
ggplot object
}
\description{
For binary EV
}
