% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_differneces}
\alias{plot_differneces}
\title{Plot differences in beta by iteration}
\usage{
plot_differneces(diffs)
}
\arguments{
\item{diffs}{}
}
\value{
ggplot object
}
\description{
Plot differences in beta by iteration
}
