% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_initial_fit}
\alias{plot_initial_fit}
\title{plot_initial_fit}
\usage{
plot_initial_fit(res, K)
}
\arguments{
\item{res}{Cleverly model object}

\item{K}{Number of responses}

\item{gammas}{Vector of dimension L + 1 for penalizing the D matrix}
}
\value{
ggplot object
}
\description{
plot_initial_fit
}
