% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverly_visualizations.R
\name{plot_r_convergence}
\alias{plot_r_convergence}
\title{plot_r_convergence}
\usage{
plot_r_convergence(res)
}
\arguments{
\item{res}{Cleverly model object}
}
\value{
ggplot
}
\description{
plot_r_convergence
}
