% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{sim_noZ}
\alias{sim_noZ}
\title{Use Chenyangs setup to simulate count data wtih 3 clusters}
\usage{
sim_noZ(
  n = 20,
  range_start = 5000,
  range_end = 20000,
  nknots = 3,
  order = 3,
  user_var = 1000,
  structure = 1,
  al = 0.4
)
}
\value{
data Matrix with columns time, individual, capture number, totaln, counts
}
\description{
No external variables.
}
