% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{simulation_data}
\alias{simulation_data}
\title{Simulate data}
\usage{
simulation_data(
  n = 20,
  range_start = 5000,
  range_end = 20000,
  nknots = 3,
  K = 12,
  order = 3,
  user_var = 1000,
  cor_str,
  Z_type = "binary",
  rho = 0.4,
  miss_p = 0.6,
  prob1 = 0.4,
  maxt = 1,
  baseline_fxns = list(function(t) 2 * cos(2 * pi * t), function(t) 2 * cos(2 * pi * t),
    function(t) 2 * cos(2 * pi * t), function(t) 2 * cos(2 * pi * t), function(t) 1 - 2 *
    exp(-6 * t), function(t) 1 - 2 * exp(-6 * t), function(t) 1 - 2 * exp(-6 * t),
    function(t) 1 - 2 * exp(-6 * t), function(t) -3 * t + 3, function(t) -3 * t + 3,
    function(t) -3 * t + 3, function(t) -3 * t + 3),
  slope_fxns = list(function(t) sqrt(t), function(t) 2 * t, function(t) -3 * t + 3,
    function(t) 4 * t, function(t) -5 * t + 5, function(t) -2 * t, function(t) 3 * t,
    function(t) log1p(t), function(t) -4 * t + 1, function(t) 5 * t, function(t) 2 * t^2,
    function(t) sin(pi * t))
)
}
\arguments{
\item{n}{Number of subjects}

\item{range_start}{Starting count range for the Dirichlet multinomial}

\item{range_end}{Ending count range for the Dirichlet multinomial}

\item{nknots}{Number of knots}

\item{K}{Number of responses}

\item{order}{Order of the B-spline basis}

\item{user_var}{User supplied variance}

\item{cor_str}{Type of correlation structure (IND, CON, AR1)}

\item{rho}{Correlation parameter}

\item{miss_p}{proportion of missing samples}

\item{prob1}{Probability of the binary external variable being 1}

\item{slope_base}{type of slope/intercept clustering curves to generate}
}
\value{
data Matrix with columns time, individual, capture number, totaln, counts
}
\description{
Use Chenyang's setup to simulate count data with 3 clusters,
one external variable binary with prob prob1 of being 1
}
