% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm3.R
\name{update_lambda}
\alias{update_lambda}
\title{Update lambda step of the admm algorithm}
\usage{
update_lambda(beta_lp, v, lambda, A, theta)
}
\arguments{
\item{theta}{}
}
\description{
Update lambda step of the admm algorithm
}
